tweak structure a bit
