+++ src/reqwest/src/async.rs
@@ -1,107 +1,5 @@
 /*! Asynchronous http client. */
 
-/*
-TODO: Figure out how connection management should work.
-
-We could do something like:
-
-- Remove base_url` from `RequestParams`
-- Add trait for `AsyncConnector` (or whatever)
-- Implement for `String`, `&'static str`, `SocketAddrs`, whatever
-- Add a new client for sniffing connections if necessary from a base addr (wraps another client)
-
-```
-let res = cli.elastic_req(params);
-let res = cli.elastic_req(params.connect_with("http://localhost:9200"));
-```
-
-Or it could be:
-
-```
-Connector: RequestParams -> RequestParams (possibly different url)?
-```
-
-So it would end up looking like:
-
-```
-let client = ReqwestClient::new();
-let connector = ClusterSniffer::new(client.clone(), RequestParams::new("baseaddress"));
-
-let params_future = connector.next();
-
-params_future.and_then(|params| {
-    client.elastic_req(params, SimpleSearchRequest::new())
-});
-```
-
-How would this work for a static list of addresses?
-
-It really is a separate concern to the rest of the parameters...
-We could bake multiple endpoints in and require a strategy to pick it?
-
-What's the optimal API?
-
-One where things _just work_, but you've got the freedom to tweak it if you want.
-Supplying a single address is convenient.
-Supplying multiple addresses is also convenient.
-Sniffing the cluster state is useful.
-
-It needs to be future-proof, so we need some way to bolt additional features on to these things.
-
-We could add a super-trait that captures some of these things and defaults whatever is missing:
-
-```
-trait GetRequestParams {
-    fn url(&self) -> Url;
-    fn headers(&self) -> Headers;
-}
-
-impl GetRequestParams for RequestParams {
-    fn url(&self) -> Url {
-        // round robin?
-    }
-}
-```
-
-Nah. I don't like it.
-
-So what if we _do_ require a way to get a url for each request?
-
-```
-let res = cli.elastic_req("http://localhost:9200", &params, SimpleSearchRequest::new());
-```
-
-Seems like it could lead to surprising  behaviour.
-From a purity point of view I like keeping `RequestParams` focused on a _single_ request.
-All the things here will end up _on_ the request.
-
-So how do multiple addresses come into this then?
-It seems like an extension in this case.
-But it doesn't make _any_ sense to have base request params with a single url if you're going to have multiple.
-Maybe just a `Fn(Url) -> RequestParams`?
-
-```
-// MultipleAddresses<TStrategy, TFactory>
-// impl TStrategy for RoundRobin
-// impl TFactory for Fn(Url) -> RequestParams
-let client = ReqwestClient::new();
-let connector = MultipleAddresses::round_robin(&["http://a:9200", "http://b:9200"], |addr| RequestParams::new(addr));
-
-let params_future = connector.next();
-
-params_future.and_then(|params| {
-    client.elastic_req(params, SimpleSearchRequest::new())
-});
-```
-
-How could we carry this around in `elastic`?
-Should it be possible to set the sniffer for a given request?
-And are we just going to have to box it to keep things ergonomic?
-It could just be an enum, but I don't think a `Box` is going to hurt anyone, really.
-
-So I'm happy with the _general idea_ of this.
-*/
-
 use std::mem;
 use std::ops::Deref;
 use bytes::Bytes;
